rename files, move types to types folder
